fix(trusty-mpm): split decommission.rs test module, clear pending_decision on delete_record
Code-critic review on PR #4430 flagged two structural issues:

- decommission.rs grew to 528 SLOC after the #4400 fix, tripping CI's
  500-line production-file cap. Extract its inline `mod tests` into a
  sibling `decommission_tests.rs`, following the exact precedent
  dedup.rs/dedup_tests.rs and decommission_worktree_tests.rs already
  established -- production logic stays under PROD_CAP, tests move to
  a sibling under the more generous TEST_CAP. Pure relocation, no
  behavior change.
- delete_record (the Deleted soft-delete path) transitioned to a
  terminal state without clearing pending_decision/proposed_default,
  same bug class as decommission_with_root. Clear both there too.

Closes #4400
